Remove the platforms subsystem.
This feature could be useful but we never actually need to apply
distribution-specific patches: that's the job of each package's build system. In
candi this feature is only used to work around some parmetis versioning problems
with old versions of CentOS - this is simply not applicable to us in any way.

The downside to keeping this patch is that it is very pedantic: we need a text
file for each supported distribution, even though that text file will be
empty (or contain just comments). Hence, things don't work on arbitrary
distributions for no good reason - not a good situation.

We can always revert this patch later if we want to use this subsystem again.