Fix problems with SAMRAI + gcc-15.
The root cause is that SAMRAI looks for DBL_NAN etc., which were macros
defined in pre-standardized C++, with the C preprocessor: this detects
them in C23 mode but not in C++ mode, which results in compilation
errors.

None of this has been necessary since C++98, so just use
std::numeric_limits instead.